+++ lib/fetchHttpClient.ts
@@ -139,6 +139,7 @@ export abstract class FetchHttpClient implements HttpClient {
       headers: httpRequest.headers.rawHeaders(),
       method: httpRequest.method,
       signal: abortController.signal,
+      redirect: "manual",
       ...platformSpecificRequestInit,
     };
 
@@ -155,6 +156,8 @@ export abstract class FetchHttpClient implements HttpClient {
           ? ((response.body as unknown) as NodeJS.ReadableStream)
           : undefined,
         bodyAsText: !httpRequest.streamResponseBody ? await response.text() : undefined,
+        redirected: response.redirected,
+        url: response.url,
       };
 
       const onDownloadProgress = httpRequest.onDownloadProgress;